% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.bib.R
\name{check.bib}
\alias{check.bib}
\title{Check Bibliography}
\usage{
check.bib(bibfile)
}
\arguments{
\item{bibfile}{File path to .bib file}
}
\value{
Returns a list of warnings, if any articles in .bib file have any associated retractions, erratums or expressions of concern. If it does not, returns an empty list.
}
\description{
\code{check.bib} searches for all articles from a .bib file on Pubmed, and returns any associated retractions, erratums or expressions of concern.
}
