% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.crossref.R
\name{check.crossref}
\alias{check.crossref}
\title{Check CrossRef}
\usage{
check.crossref(articlename)
}
\arguments{
\item{articlename}{Title of article (in quotes)}
}
\value{
Returns a list of warnings, if article has any associated retractions.
}
\description{
\code{check.crossref} searches for a single article on Pubmed, and returns any associated retractions, erratums or expressions of concern.
}
\examples{

check.crossref("Ileal-lymphoid-nodular hyperplasia, non-specific colitis, and pervasive developmental disorder in children")
# (returns warning that article may have been retracted)
}
